Good advancement problem of locating where to put the code lines

+++ Stadium-Managing/src/main/java/simu/framework/Clock.java
@@ -16,7 +16,7 @@ public class Clock {
         this.clock = clock;
     }
 
-    public int getClock() {
-        return (int) clock;
+    public double getClock() {
+        return clock;
     }
 }
